requirements.txt
pinecone
google-generativeai
uvicorn[standard]
fastapi
